requirements.txt
aiohttp==3.6.2
motor==2.1.0
pydantic==1.4
pyyaml==5.3.1